requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.1
backoff==1.11.1
beautifulsoup4==4.11.1
bottle==0.12.25
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
clr-loader==0.2.5
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.6
cryptography==39.0.1
cssselect==1.2.0
cycler==0.11.0
databricks-cli==0.17.3
dill==0.3.6
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.7
entrypoints==0.4
exceptiongroup==1.1.0
farm-haystack==1.13.2
filelock==3.8.2
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
h11==0.14.0
huggingface-hub==0.11.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
inflect==6.0.2
isodate==0.6.1
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
llvmlite==0.39.1
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mlflow==2.0.1
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
multidict==6.0.4
networkx==2.8.8
nltk==3.7
num2words==0.5.12
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.26.5
outcome==1.2.0
packaging==21.3
pandas==1.5.2
parsel==1.7.0
pdf2image==1.16.1
Pillow==9.4.0
posthog==2.2.0
Protego==0.2.1
protobuf==3.20.2
proxy-tools==0.1.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyDispatcher==2.0.6
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
pythonnet==3.0.1
pytz==2022.6
pywebview==4.0.2
pywin32==305
PyYAML==6.0
quantulum3==0.7.11
querystring-parser==1.2.4
queuelib==1.6.2
rank-bm25==0.2.2
rapidfuzz==2.7.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
service-identity==21.1.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
tabulate==0.9.0
threadpoolctl==3.1.0
tika==1.24
tkhtmlview==0.1.3
tldextract==3.4.0
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
waitress==2.1.2
webdriver-manager==3.8.5
websocket-client==1.4.2
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
zipp==3.11.0
zope.interface==5.5.2
